pyproject.toml
[tool.poetry]
name = "mini-telegram"
version = "0.1.0"
description = ""
authors = ["macbook_uz <fakhriddinu@icloud.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.2"
uvicorn = "^0.32.0"
sqlalchemy = "^2.0.36"
jinja2 = "^3.1.4"
asyncpg = "^0.30.0"
psycopg2-binary = "^2.9.10"
bcrypt = "^4.2.0"
fastapi-storages = "^0.3.0"
greenlet = "^3.1.1"
python-multipart = "^0.0.12"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = "^3.3.0"
pyjwt = "^2.9.0"
itsdangerous = "^2.2.0"
wsproto = "^1.2.0"
aiogram = "^3.13.1"
python-dotenv = "^1.0.1"
celery = {extras = ["redis"], version = "^5.4.0"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
